fix: prompting heuristics, selection timers, allocation final round, SQL
- Heuristics: cumulative word count is order-independent; T_NA/T_HU per-image
  line matches round total words at that prompt (show_prompting_heuristics).
- app: ensure DB round row before prompts; synthetic negative prompt_id when
  DB row missing; clearer post-survey env hint; skip image_selections DB row
  for synthetic ids; _persist uses cumulative HU snapshot where applicable.
- game.js: visible error when image gen fails; guard timer_update to game screen;
  clear selection-timer interval before restarts to stop flicker.
- ballot_balancer: final round works with 2+ players (repeat peers when needed);
  round-3 selected_images still required for owner pool.
- SQL: FK types integer for games/prompts; pre_survey RLS policy + bootstrap note.
- CSS: .image-gen-error-msg for failed-generation placeholder.

Made-with: Cursor

+++ sql/experiment_voting_schema.sql
@@ -28,7 +28,8 @@ ALTER TABLE image_selections ADD COLUMN IF NOT EXISTS selection_steps_back_from_
 -- ---------------------------------------------------------------------------
 CREATE TABLE IF NOT EXISTS voting_rounds (
   voting_round_id   bigint GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
-  game_id           bigint NOT NULL REFERENCES games(game_id) ON DELETE CASCADE,
+  -- INTEGER matches games.game_id (SERIAL) from bootstrap schema
+  game_id           integer NOT NULL REFERENCES games(game_id) ON DELETE CASCADE,
   voting_round_index int NOT NULL,
   kind              text NOT NULL CHECK (kind IN ('hardcoded', 'final')),
   fixture_set_key   text,
@@ -45,7 +46,7 @@ CREATE INDEX IF NOT EXISTS idx_voting_rounds_game ON voting_rounds(game_id);
 -- ---------------------------------------------------------------------------
 CREATE TABLE IF NOT EXISTS voter_ballots (
   ballot_id         bigint GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
-  game_id           bigint NOT NULL REFERENCES games(game_id) ON DELETE CASCADE,
+  game_id           integer NOT NULL REFERENCES games(game_id) ON DELETE CASCADE,
   voting_round_id   bigint NOT NULL REFERENCES voting_rounds(voting_round_id) ON DELETE CASCADE,
   voter_player_id   text NOT NULL REFERENCES players(player_id) ON DELETE CASCADE,
   UNIQUE (voting_round_id, voter_player_id)
@@ -64,7 +65,7 @@ CREATE TABLE IF NOT EXISTS ballot_options (
   fixture_image_id    text,
   image_url           text NOT NULL,
   owner_player_id     text REFERENCES players(player_id) ON DELETE SET NULL,
-  prompt_id           bigint REFERENCES prompts(prompt_id) ON DELETE SET NULL,
+  prompt_id           integer REFERENCES prompts(prompt_id) ON DELETE SET NULL,
   heuristic_snapshot  jsonb,
   UNIQUE (ballot_id, slot_index)
 );
@@ -77,7 +78,7 @@ CREATE INDEX IF NOT EXISTS idx_ballot_options_ballot ON ballot_options(ballot_id
 CREATE TABLE IF NOT EXISTS point_allocations (
   id                  bigint GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
   ballot_id           bigint NOT NULL UNIQUE REFERENCES voter_ballots(ballot_id) ON DELETE CASCADE,
-  game_id             bigint NOT NULL REFERENCES games(game_id) ON DELETE CASCADE,
+  game_id             integer NOT NULL REFERENCES games(game_id) ON DELETE CASCADE,
   voting_round_id     bigint NOT NULL REFERENCES voting_rounds(voting_round_id) ON DELETE CASCADE,
   voter_player_id     text NOT NULL REFERENCES players(player_id) ON DELETE CASCADE,
   points_slot_1       int NOT NULL CHECK (points_slot_1 >= 0 AND points_slot_1 <= 10),
